Move functions.py to sans.py
There is no reason to have an entirely seperate file for one function (i
mean there's not really a point having this make_embed function anyway
but that's a different issue)
