+++ sans.py
@@ -6,7 +6,22 @@ import os
 import discord
 from discord.ext import commands
 
-from functions import make_embed
+def make_embed(title=None, description=None, author=None, thumbnail=None, url=None, color=None, image=None, name=None) -> discord.Embed: 
+    """Cleaner way to make embeds
+    Thumbnail and author are taken as kwargs, title is default "Guide" """ 
+    if isinstance(color, int):
+        color = discord.Color(color)
+    
+    embed = discord.Embed(
+        title=title,
+        color=color,
+        url=url,
+        description=description
+    )
+    if name is not None:   embed.set_author(name=name)
+    if image is not None:   embed.set_image(url=image)
+    if thumbnail is not None:   embed.set_thumbnail(url=thumbnail)
+    return embed
 
 bot = commands.Bot(command_prefix = '.')
